fix: 개발을 위한 토큰 만료시간 재수정

+++ src/main/java/com/jungle/Tabbit/global/config/security/jwt/JwtConst.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum JwtConst {
-    ACCESS_TOKEN_EXPIRE_TIME(Long.MAX_VALUE),
+    ACCESS_TOKEN_EXPIRE_TIME(Integer.MAX_VALUE),
     REFRESH_TOKEN_EXPIRE_TIME(3600000 * 24);
 
     private final long time;
